fix: create typeRoots for global types
