+++ packages/client/polyfills.ts
@@ -1,5 +1,3 @@
-import './types/modules.d'
-
 if (__PRODUCTION__) {
   __webpack_public_path__ = window.__ACTION__.publicPath
 }
