--- packages/client/serviceWorker/modules.d.ts
@@ -1,3 +1,0 @@
-// this is just to get typescript to stop complaining about imports
-declare const __PRODUCTION__: string
-declare const __APP_VERSION__: string
