--- packages/client/types/reactHTML4.d.ts
@@ -1,14 +1,0 @@
-import 'react'
-declare module 'react' {
-  interface TdHTMLAttributes<_T> {
-    height?: string | number
-    width?: string | number
-    bgcolor?: string
-  }
-  interface TableHTMLAttributes<_T> {
-    align?: 'center' | 'left' | 'right'
-    bgcolor?: string
-    height?: string | number
-    width?: string | number
-  }
-}
