--- packages/client/types/shared.d.ts
@@ -1,2 +1,0 @@
-// The client require things from the server package, but it does, so we need this
-import '../../server/types/modules'
