+++ packages/client/utils/relay/getGraphQLError.ts
@@ -1,21 +1,5 @@
 import type {PayloadError} from 'relay-runtime'
 
-declare module 'relay-runtime' {
-  interface PayloadError extends Error {
-    message: string
-    locations?:
-      | Array<{
-          line: number
-          column: number
-        }>
-      | undefined
-    path?: Array<string | number>
-    extensions?: {
-      code?: string
-    }
-  }
-}
-
 const getServerError = (errors: ReadonlyArray<PayloadError> | null | undefined) =>
   (errors && errors[0]) || undefined
 const getPayloadError = (res: {[key: string]: any} | null) => {
