--- packages/embedder/types/modules.d.ts
@@ -1,4 +1,0 @@
-declare module 'franc-min' {
-  import f from 'franc'
-  export = f
-}
