--- packages/embedder/types/shared.d.ts
@@ -1,1 +1,0 @@
-import '../../server/types/modules'
