+++ packages/server/types/custom.d.ts
@@ -1,19 +1,5 @@
-import 'react'
 import type ScheduledJobMeetingStageTimeLimit from '../database/types/ScheduledJobMetingStageTimeLimit'
 import type ScheduledTeamLimitsJob from '../database/types/ScheduledTeamLimitsJob'
-declare module 'react' {
-  interface TdHTMLAttributes<_T> {
-    height?: string | number
-    width?: string | number
-    bgcolor?: string
-  }
-  interface TableHTMLAttributes<_T> {
-    align?: 'center' | 'left' | 'right'
-    bgcolor?: string
-    height?: string | number
-    width?: string | number
-  }
-}
 
 export interface OAuth2Success {
   access_token: string
